requirements.txt
requests
allure-pytest
